requirements.txt
flask>=2.3.0
openai-whisper>=20231117
anthropic>=0.25.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
numpy>=1.24.0
